Refactor parse objects into separate files
