+++ src/parse/DataScalar.ts
@@ -1,0 +1,15 @@
+import { GraphQLScalarType } from 'graphql';
+
+export const DataScalar = new GraphQLScalarType({
+    name: 'Data',
+    description: 'Any data',
+    parseValue(value: any) {
+        return value;
+    },
+    serialize(value: any) {
+        return value;
+    },
+    parseLiteral(ast) {
+        return ast
+    }
+})
