+++ src/parse/MetaData.ts
@@ -1,0 +1,12 @@
+import { ObjectType, Field } from 'type-graphql'
+import { IMetaData } from 'corewar'
+
+@ObjectType()
+export class MetaData implements IMetaData {
+    @Field()
+    author!: string
+    @Field()
+    name!: string
+    @Field()
+    strategy!: string
+}
