+++ src/parse/Token.ts
@@ -1,0 +1,13 @@
+import { ObjectType, Field } from 'type-graphql'
+import { IToken, TokenCategory } from 'corewar'
+import { Position } from './Position'
+
+@ObjectType()
+export class Token implements IToken {
+    @Field()
+    category!: TokenCategory
+    @Field()
+    lexeme!: string
+    @Field()
+    position!: Position
+}
